refactor: gas utils. add separate token price
