+++ src/lib/hooks/use-best-quote/utils/issuance/legacy-quote.ts
@@ -5,10 +5,10 @@ import { Issuance } from '@/app/legacy/config'
 import { LeveragedRethStakingYield } from '@/app/legacy/config/tokens/mainnet'
 import { POLYGON } from '@/constants/chains'
 import { RETH, type Token } from '@/constants/tokens'
-import { getTokenPrice } from '@/lib/hooks/use-token-price'
 import { formatWei, isSameAddress } from '@/lib/utils'
 import { getFullCostsInUsd } from '@/lib/utils/costs'
 import { getGasLimit } from '@/lib/utils/gas'
+import { getTokenPrice } from '@/lib/utils/token-price'
 
 import { type Quote, type QuoteTransaction, QuoteType } from '../../types'
 
@@ -97,7 +97,11 @@ export async function getLegacyRedemptionQuote(
     }
 
     const defaultGasEstimate = BigInt(200_000)
-    const { ethPrice, gas } = await getGasLimit(transaction, defaultGasEstimate)
+    const { ethPrice, gas } = await getGasLimit(
+      transaction,
+      defaultGasEstimate,
+      publicClient,
+    )
     transaction.gas = gas.limit
 
     const inputTokenAmountUsd =
